+++ test/integration/monitor_valid_test.go
@@ -1,0 +1,31 @@
+package integration
+
+import (
+	"testing"
+	"time"
+)
+
+func TestMonitorWithValidCheckpoint(t *testing.T) {
+	ctx, binary, checkpointFile, monitorPort, mockServer := setupTest(t)
+	defer mockServer.Close()
+
+	runCmd, logs := startMonitor(t, ctx, binary, checkpointFile, monitorPort, mockServer)
+
+	metrics := fetchMetrics(t, monitorPort)
+	validateLogsAndMetrics(t, logs, metrics, MonitorExpectations{
+		ExpectErrorLog:       false,
+		ExpectedFailureCount: 0,
+		ExpectedTotalCount:   1,
+	})
+
+	time.Sleep(1 * time.Second)
+
+	metrics = fetchMetrics(t, monitorPort)
+	validateLogsAndMetrics(t, logs, metrics, MonitorExpectations{
+		ExpectErrorLog:       false,
+		ExpectedFailureCount: 0,
+		ExpectedTotalCount:   2,
+	})
+
+	stopMonitor(t, runCmd)
+}
